+++ gauguin-core/src/main/kotlin/org/piepmeyer/gauguin/Theme.kt
@@ -4,4 +4,5 @@ enum class Theme {
     LIGHT,
     DARK,
     SYSTEM_DEFAULT,
+    DYNAMIC_COLORS,
 }