+++ CHANGELOG.md
@@ -10,6 +10,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 
 - Added all difficulty levels of 4x4 and 5x5 grids.
+- Added support of dynamic colors.
 
 ### Changed